+++ packages/components/src/Badge/BadgeComposition/BadgeCategory/BadgeCategory.component.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import TooltipTrigger from '../../../TooltipTrigger';
-import badgeCssModule from '../../Badge.scss';
+import badgeCssModule from '../../Badge.module.scss';
 import { getTheme } from '../../../theme';
 
 const theme = getTheme(badgeCssModule);