+++ packages/components/src/Enumeration/Header/HeaderInput.component.js
@@ -3,7 +3,7 @@ import React from 'react';
 import classNames from 'classnames';
 
 import Action from '../../Actions/Action';
-import theme from './Header.scss';
+import theme from './Header.module.scss';
 
 function headerClasses(headerError) {
 	return classNames(theme['tc-enumeration-header'], 'tc-enumeration-header', {