+++ packages/components/src/ResourcePicker/ResourcePicker.component.js
@@ -5,7 +5,7 @@ import { getTheme } from '../theme';
 import ResourceList from '../ResourceList';
 import { SORT_OPTIONS, ORDERS, STATE_FILTERS } from '../ResourceList/Toolbar';
 
-import cssModule from './ResourcePicker.scss';
+import cssModule from './ResourcePicker.module.scss';
 import ResourceListPropTypes from '../ResourceList/ResourceList.propTypes';
 
 const theme = getTheme(cssModule);