+++ packages/components/src/Rich/Error/RichError.component.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import Icon from '../../Icon';
 
-import theme from './RichError.scss';
+import theme from './RichError.module.scss';
 
 export default function RichError(props) {
 	return (