+++ packages/components/src/VirtualizedList/HeaderCheckbox/HeaderCheckbox.component.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import React, { useMemo } from 'react';
 import classnames from 'classnames';
 import Checkbox from '../../Checkbox';
-import theme from './HeaderCheckbox.scss';
+import theme from './HeaderCheckbox.module.scss';
 import getDefaultT from '../../translate';
 
 /**