+++ packages/dataviz/src/components/BarChart/TooltipCursor/TooltipCursor.component.tsx
@@ -1,6 +1,6 @@
 import React, { SVGProps } from 'react';
 import { Customized } from 'recharts';
-import styles from './TooltipCursor.component.scss';
+import styles from './TooltipCursor.component.module.scss';
 
 export type TooltipCursorProps = SVGProps<SVGElement> & {
 	dataFeature?: string;