+++ packages/dataviz/src/components/RangeFilter/handlers/DateTimeRangeHandler/DateTimeRangeHandler.tsx
@@ -6,7 +6,7 @@ import parseISO from 'date-fns/parseISO';
 import { InputDateTimePicker } from '@talend/react-components';
 import { formatDate, formatDateTime } from '../../../../formatters/formatters';
 import useRangeInputField, { InputFieldProps } from '../useRangeInputField.hook';
-import styles from './DateTimeInputField.component.scss';
+import styles from './DateTimeInputField.component.module.scss';
 import { RangeHandler } from '../range-handler.types';
 import { formatTimeTicks } from '../slider-ticks.utils';