+++ packages/forms/src/UIForm/fieldsets/Array/DefaultArrayTemplate.component.js
@@ -8,7 +8,7 @@ import Message from '../../Message';
 import { generateDescriptionId, generateErrorId } from '../../Message/generateId';
 import { I18N_DOMAIN_FORMS } from '../../../constants';
 
-import theme from './DefaultArrayTemplate.scss';
+import theme from './DefaultArrayTemplate.module.scss';
 
 function DefaultArrayTemplate(props) {
 	const { t } = useTranslation(I18N_DOMAIN_FORMS);